# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeJsonRequest": {
        "additionalProperties": false,
        "description": "JSON form of `POST /api/v1/dr/analyze` (the brief's `{ runbookMarkdown, inventory? }`).",
        "properties": {
          "inventory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SystemInventory"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "runbookMarkdown": {
            "description": "The runbook as Markdown text.",
            "minLength": 1,
            "title": "Runbookmarkdown",
            "type": "string"
          },
          "runbookName": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Label for report meta.runbookFile.",
            "title": "Runbookname"
          }
        },
        "required": [
          "runbookMarkdown"
        ],
        "title": "AnalyzeJsonRequest",
        "type": "object"
      },
...
      "DRReadinessReport": {
        "additionalProperties": false,
        "properties": {
          "aiAnalysisAvailable": {
            "default": true,
            "title": "Aianalysisavailable",
            "type": "boolean"
          },
          "aiNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ainote"
          },
          "dependencyHealth": {
            "items": {
              "$ref": "#/components/schemas/DependencyHealth"
            },
            "title": "Dependencyhealth",
            "type": "array"
          },
          "executionPlan": {
            "items": {
              "$ref": "#/components/schemas/ExecutionPhase"
            },
            "title": "Executionplan",
            "type": "array"
          },
          "gapAnalysis": {
            "items": {
              "$ref": "#/components/schemas/Gap"
            },
            "title": "Gapanalysis",
            "type": "array"
          },
          "historicalInsights": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceHistory"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ReportMeta"
          },
          "riskLevel": {
            "$ref": "#/components/schemas/RiskLevel",
            "readOnly": true
          },
          "riskScore": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Riskscore",
            "type": "integer"
          },
          "rtoAnalysis": {
            "$ref": "#/components/schemas/RtoAnalysis"
          },
          "serviceSummary": {
            "$ref": "#/components/schemas/ServiceSummary"
          },
          "singlePointsOfFailure": {
            "items": {
              "$ref": "#/components/schemas/SinglePointOfFailure"
            },
            "title": "Singlepointsoffailure",
            "type": "array"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/Suggestion"
            },
            "title": "Suggestions",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "meta",
          "serviceSummary",
          "riskScore",
          "rtoAnalysis",
          "summary",
          "riskLevel"
        ],
        "title": "DRReadinessReport",
        "type": "object"
      },
      "DependencyHealth": {
        "additionalProperties": false,
        "properties": {
          "actualStatus": {
            "$ref": "#/components/schemas/DependencyStatus"
          },
          "impact": {
            "description": "Which steps are affected",
            "title": "Impact",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "runbookAssumes": {
            "const": "available",
            "default": "available",
            "title": "Runbookassumes",
            "type": "string"
          }
        },
        "required": [
          "name",
          "actualStatus",
          "impact"
        ],
        "title": "DependencyHealth",
        "type": "object"
      },
      "DependencyStatus": {
        "description": "Health as seen from a runbook dependency; adds the not-in-inventory case.",
        "enum": [
          "UP",
          "DOWN",
          "UNREACHABLE",
          "NOT_IN_INVENTORY"
        ],
        "title": "DependencyStatus",
        "type": "string"
      },
      "DependencyTrend": {
        "additionalProperties": false,
        "properties": {
          "analyses": {
            "minimum": 0.0,
            "title": "Analyses",
            "type": "integer"
          },
          "down": {
            "default": 0,
            "minimum": 0.0,
            "title": "Down",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notInInventory": {
            "default": 0,
            "minimum": 0.0,
            "title": "Notininventory",
            "type": "integer"
          },
          "unreachable": {
            "default": 0,
            "minimum": 0.0,
            "title": "Unreachable",
            "type": "integer"
          },
          "up": {
            "default": 0,
            "minimum": 0.0,
            "title": "Up",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "analyses"
        ],
        "title": "DependencyTrend",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The shared error shape: `{error, code, details?}`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ExecutionOutcome": {
        "additionalProperties": false,
        "properties": {
          "elapsedMinutes": {
            "minimum": 0.0,
            "title": "Elapsedminutes",
            "type": "number"
          },
          "executionId": {
            "title": "Executionid",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "title": "Startedat",
            "type": "string"
          },
          "state": {
            "enum": [
              "COMPLETED",
              "FAILED",
              "ABORTED"
            ],
            "title": "State",
            "type": "string"
          },
          "statedRtoMinutes": {
            "exclusiveMinimum": 0.0,
            "title": "Statedrtominutes",
            "type": "number"
          }
        },
        "required": [
          "executionId",
          "state",
          "startedAt",
          "elapsedMinutes",
          "statedRtoMinutes"
        ],
        "title": "ExecutionOutcome",
        "type": "object"
      },
      "ExecutionPhase": {
        "additionalProperties": false,
        "properties": {
          "estimatedMinutes": {
            "minimum": 0.0,
            "title": "Estimatedminutes",
            "type": "number"
          },
          "gate": {
            "description": "What must be true before proceeding",
            "title": "Gate",
            "type": "string"
          },
          "phase": {
            "minimum": 1.0,
            "title": "Phase",
            "type": "integer"
          },
          "steps": {
            "description": "Steps that can run in parallel",
            "items": {
              "type": "integer"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "phase",
          "steps",
          "estimatedMinutes",
          "gate"
        ],
        "title": "ExecutionPhase",
        "type": "object"
      },
...
      "Gap": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "type": {
            "$ref": "#/components/schemas/GapType"
          }
        },
        "required": [
          "type",
          "description",
          "severity",
          "recommendation"
        ],
        "title": "Gap",
        "type": "object"
      },
      "GapType": {
        "enum": [
          "MISSING_STEP",
          "VAGUE_INSTRUCTION",
          "NO_VALIDATION",
          "MISSING_ROLLBACK",
          "UNVERIFIED_DEPENDENCY",
          "OWNER_AMBIGUITY",
          "HISTORICAL"
        ],
        "title": "GapType",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobView": {
        "additionalProperties": false,
        "properties": {
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "historySaved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Whether the report was stored (null while history is off)",
            "title": "Historysaved"
          },
          "jobId": {
            "title": "Jobid",
            "type": "string"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DRReadinessReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/JobState"
          }
        },
        "required": [
          "jobId",
          "status",
          "createdAt"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "ReportMeta": {
        "additionalProperties": false,
        "properties": {
          "agentVersion": {
            "title": "Agentversion",
            "type": "string"
          },
          "analysisTimeMs": {
            "minimum": 0.0,
            "title": "Analysistimems",
            "type": "number"
          },
          "analyzedAt": {
            "format": "date-time",
            "title": "Analyzedat",
            "type": "string"
          },
          "inventoryFile": {
            "title": "Inventoryfile",
            "type": "string"
          },
          "runbookFile": {
            "title": "Runbookfile",
            "type": "string"
          }
        },
        "required": [
          "analyzedAt",
          "runbookFile",
          "inventoryFile",
          "agentVersion",
          "analysisTimeMs"
        ],
        "title": "ReportMeta",
        "type": "object"
      },
...
      "RiskLevel": {
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
      "RtoAnalysis": {
        "additionalProperties": false,
        "properties": {
          "bottleneckSteps": {
            "items": {
              "type": "integer"
            },
            "title": "Bottlenecksteps",
            "type": "array"
          },
          "bufferMinutes": {
            "description": "Stated RTO minus estimated total; may be negative",
            "title": "Bufferminutes",
            "type": "number"
          },
          "feasible": {
            "title": "Feasible",
            "type": "boolean"
          },
          "statedRtoMinutes": {
            "exclusiveMinimum": 0.0,
            "title": "Statedrtominutes",
            "type": "number"
          },
          "totalEstimatedMinutes": {
            "minimum": 0.0,
            "title": "Totalestimatedminutes",
            "type": "number"
          }
        },
        "required": [
          "feasible",
          "totalEstimatedMinutes",
          "statedRtoMinutes",
          "bufferMinutes"
        ],
        "title": "RtoAnalysis",
        "type": "object"
      },
      "SampleFile": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "content"
        ],
        "title": "SampleFile",
        "type": "object"
      },
...
      "ServiceHistory": {
        "additionalProperties": false,
        "description": "Facts computed by code from stored analyses and live executions only.",
        "properties": {
          "analysesConsidered": {
            "minimum": 0.0,
            "title": "Analysesconsidered",
            "type": "integer"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyTrend"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "dryRuns": {
            "description": "Counted but never measured",
            "minimum": 0.0,
            "title": "Dryruns",
            "type": "integer"
          },
          "executions": {
            "description": "Newest first",
            "items": {
              "$ref": "#/components/schemas/ExecutionOutcome"
            },
            "title": "Executions",
            "type": "array"
          },
          "liveRuns": {
            "minimum": 0.0,
            "title": "Liveruns",
            "type": "integer"
          },
          "riskScores": {
            "description": "Oldest first",
            "items": {
              "type": "integer"
            },
            "title": "Riskscores",
            "type": "array"
          },
          "serviceName": {
            "title": "Servicename",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepHistory"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "serviceName",
          "analysesConsidered",
          "liveRuns",
          "dryRuns"
        ],
        "title": "ServiceHistory",
        "type": "object"
      },
      "ServiceSummary": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "statedRPO": {
            "description": "Minutes",
            "exclusiveMinimum": 0.0,
            "title": "Statedrpo",
            "type": "number"
          },
          "statedRTO": {
            "description": "Minutes",
            "exclusiveMinimum": 0.0,
            "title": "Statedrto",
            "type": "number"
          }
        },
        "required": [
          "name",
          "owner",
          "statedRTO",
          "statedRPO"
        ],
        "title": "ServiceSummary",
        "type": "object"
      },
      "Severity": {
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "Severity",
        "type": "string"
      },
      "SinglePointOfFailure": {
        "additionalProperties": false,
        "properties": {
          "affectedSteps": {
            "items": {
              "type": "integer"
            },
            "title": "Affectedsteps",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "mitigationSuggestion": {
            "title": "Mitigationsuggestion",
            "type": "string"
          }
        },
        "required": [
          "description",
          "mitigationSuggestion"
        ],
        "title": "SinglePointOfFailure",
        "type": "object"
      },
...
      "StepHistory": {
        "additionalProperties": false,
        "properties": {
          "estimatedMinutes": {
            "minimum": 0.0,
            "title": "Estimatedminutes",
            "type": "number"
          },
          "failed": {
            "default": 0,
            "minimum": 0.0,
            "title": "Failed",
            "type": "integer"
          },
          "fingerprint": {
            "description": "SHA-256 of the normalized action and target system",
            "title": "Fingerprint",
            "type": "string"
          },
          "liveRuns": {
            "minimum": 0.0,
            "title": "Liveruns",
            "type": "integer"
          },
          "manualDone": {
            "default": 0,
            "minimum": 0.0,
            "title": "Manualdone",
            "type": "integer"
          },
          "maxActiveMinutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxactiveminutes"
          },
          "medianActiveMinutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Running (or manual) until the step's final state",
            "title": "Medianactiveminutes"
          },
          "medianElapsedMinutes": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Including approval waits after the run started",
            "title": "Medianelapsedminutes"
          },
          "retries": {
            "default": 0,
            "minimum": 0.0,
            "title": "Retries",
            "type": "integer"
          },
          "rolledBack": {
            "default": 0,
            "minimum": 0.0,
            "title": "Rolledback",
            "type": "integer"
          },
          "skipped": {
            "default": 0,
            "minimum": 0.0,
            "title": "Skipped",
            "type": "integer"
          },
          "stepNumber": {
            "description": "In the runbook it was matched to",
            "minimum": 1.0,
            "title": "Stepnumber",
            "type": "integer"
          },
          "succeeded": {
            "default": 0,
            "minimum": 0.0,
            "title": "Succeeded",
            "type": "integer"
          },
          "targetSystem": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetsystem"
          },
          "unknown": {
            "default": 0,
            "minimum": 0.0,
            "title": "Unknown",
            "type": "integer"
          }
        },
        "required": [
          "stepNumber",
          "fingerprint",
          "estimatedMinutes",
          "liveRuns"
        ],
        "title": "StepHistory",
        "type": "object"
      },
...
      "Suggestion": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "priority": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Priority",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "priority",
          "title",
          "detail"
        ],
        "title": "Suggestion",
        "type": "object"
      },
      "SystemInventory": {
        "additionalProperties": false,
        "properties": {
          "services": {
            "items": {
              "$ref": "#/components/schemas/InventoryService"
            },
            "title": "Services",
            "type": "array"
          }
        },
        "title": "SystemInventory",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Parse DR runbooks, check dependency health and produce readiness reports.",
    "title": "DR Readiness Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dr/analyze": {
      "get": {
        "description": "Analyze server-side files, given as paths relative to `API_ALLOWED_DIR`.",
        "operationId": "get_analyze_api_v1_dr_analyze_get",
        "parameters": [
          {
            "in": "query",
            "name": "runbook",
            "required": true,
            "schema": {
              "examples": [
                "runbooks/x.md"
              ],
              "maxLength": 512,
              "minLength": 1,
              "title": "Runbook",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inventory",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Inventory"
            }
          },
          {
            "description": "Wait (bounded) and return the report if done.",
            "in": "query",
            "name": "wait",
            "required": false,
            "schema": {
              "default": false,
              "description": "Wait (bounded) and return the report if done.",
              "title": "Wait",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DRReadinessReport"
                }
              }
            },
            "description": "wait=true and the job finished in time"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Job accepted; poll the Location header"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Analyze",
        "tags": [
          "analysis"
        ]
      },
      "post": {
        "description": "Analyze an uploaded runbook: multipart (`runbook` + optional `inventory` files) or JSON.",
        "operationId": "post_analyze_api_v1_dr_analyze_post",
        "parameters": [
          {
            "description": "Wait (bounded) and return the report if done.",
            "in": "query",
            "name": "wait",
            "required": false,
            "schema": {
              "default": false,
              "description": "Wait (bounded) and return the report if done.",
              "title": "Wait",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeJsonRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "inventory": {
                    "format": "binary",
                    "type": "string"
                  },
                  "runbook": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "runbook"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DRReadinessReport"
                }
              }
            },
            "description": "wait=true and the job finished in time"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Job accepted; poll the Location header"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Post Analyze",
        "tags": [
          "analysis"
        ]
      }
    },
...
    "/api/v1/dr/samples/file": {
      "get": {
        "description": "The text of one sample file (a path from `GET /api/v1/dr/samples`).",
        "operationId": "get_sample_api_v1_dr_samples_file_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "examples": [
                "runbooks/x.md"
              ],
              "maxLength": 512,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleFile"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sample",
        "tags": [
          "samples"
        ]
      }
    },
...
  }
}